refactor portfolios folder
