+++ src/portfolios/portfolio/components/FundingTracker/FundingTracker.spec.ts
@@ -2,7 +2,7 @@ import Vue from "vue";
 import Vuetify from "vuetify";
 import { createLocalVue, mount, Wrapper } from "@vue/test-utils";
 import { DefaultProps } from "vue/types/options";
-import FundingTracker from "@/portfolio/FundingTracker.vue";
+import FundingTracker from "@/portfolios/portfolio/components/FundingTracker/FundingTracker.vue";
 import { nextTick } from "process";
 
 import Chart, { ChartData } from "chart.js/auto";
