+++ src/portfolios/portfolio/components/Index.spec.ts
@@ -2,17 +2,17 @@ import Vue from "vue";
 import Vuetify from "vuetify";
 import { createLocalVue, mount, Wrapper } from "@vue/test-utils";
 import { DefaultProps } from "vue/types/options";
-import TaskOrder from "@/portfolio/TaskOrder.vue";
+import Index  from "@/portfolios/portfolio/components/Index.vue";
 Vue.use(Vuetify);
 
-describe("Testing TaskOrder Component", () => {
+describe("Testing index Component", () => {
   const localVue = createLocalVue();
   let vuetify: Vuetify;
   let wrapper: Wrapper<DefaultProps & Vue, Element>;
 
   beforeEach(() => {
     vuetify = new Vuetify();
-    wrapper = mount(TaskOrder, {
+    wrapper = mount(Index, {
       localVue,
       vuetify,
     });
