+++ src/portfolios/portfolio/components/shared/CSPCard.spec.ts
@@ -2,7 +2,7 @@ import Vue from "vue";
 import Vuetify from "vuetify";
 import { createLocalVue, mount, Wrapper } from "@vue/test-utils";
 import { DefaultProps } from "vue/types/options";
-import CSPCard from "@/portfolio/components/CSPCard.vue";
+import CSPCard from "@/portfolios/portfolio/components/shared/CSPCard.vue";
 import SlideoutPanel from "@/store/slideoutPanel";
 Vue.use(Vuetify);
