+++ src/portfolios/portfolio/components/shared/CSPCard.vue
@@ -64,9 +64,9 @@ import Vue from "vue";
 
 import { Component, Prop } from "vue-property-decorator";
 import ATATSVGIcon from "@/components/icons/ATATSVGIcon.vue";
-import { SlideoutPanelContent } from "../../../types/Global";
+import { SlideoutPanelContent } from "../../../../../types/Global";
 import SlideoutPanel from "@/store/slideoutPanel";
-import AccessingCSPLearnMore from "@/portfolio/components/AccessingCSPLearnMore.vue";
+import AccessingCSPLearnMore from "@/portfolios/portfolio/components/shared/AccessingCSPLearnMore.vue";
 @Component({
   components: {
     ATATSVGIcon,
