+++ src/portfolios/portfolio/components/shared/Members.spec.ts
@@ -2,17 +2,17 @@ import Vue from "vue";
 import Vuetify from "vuetify";
 import { createLocalVue, mount, Wrapper } from "@vue/test-utils";
 import { DefaultProps } from "vue/types/options";
-import Index  from "@/portfolio/Index.vue";
+import Members from "@/portfolios/portfolio/components/shared/Members.vue";
 Vue.use(Vuetify);
 
-describe("Testing index Component", () => {
+describe("Testing Members Component", () => {
   const localVue = createLocalVue();
   let vuetify: Vuetify;
   let wrapper: Wrapper<DefaultProps & Vue, Element>;
 
   beforeEach(() => {
     vuetify = new Vuetify();
-    wrapper = mount(Index, {
+    wrapper = mount(Members, {
       localVue,
       vuetify,
     });
